Poniendole permisos a la API de usuarios

+++ users/api.py
@@ -6,10 +6,15 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from users.serializers import UserSerializer
+from users.permissions import UserPermission
 
 
 class UserListAPI(APIView):
+
+    permission_classes = (UserPermission,)
+
     def get(self, request):
+        self.check_permissions(request)
         paginator = PageNumberPagination()
         users = User.objects.all()
         paginator.paginate_queryset(users, request)
@@ -18,6 +23,7 @@ class UserListAPI(APIView):
         return paginator.get_paginated_response(serializer_users)
 
     def post(self, request):
+        self.check_permissions(request)
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             new_user = serializer.save()
@@ -30,16 +36,20 @@ class UserDetailAPI(APIView):
     """
     Pk es la variable que me pasa la URL
     """
+    permission_classes = (UserPermission,)
 
     def get(self, request, var):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=var)
-
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         serializer_users = serializer.data
         return Response(serializer_users)
 
     def put(self, request, var):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=var)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(instance=user, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -48,6 +58,8 @@ class UserDetailAPI(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, var):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=var)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
